+++ packages/typegpu/src/data/dataIO.ts
@@ -1,5 +1,11 @@
-import type { ISerialInput, ISerialOutput } from 'typed-binary';
-import type { Infer, InferRecord } from '../shared/repr.ts';
+import {
+  BufferReader,
+  BufferWriter,
+  getSystemEndianness,
+  type ISerialInput,
+  type ISerialOutput,
+} from 'typed-binary';
+import type { Infer, InferInput, InferRecord } from '../shared/repr.ts';
 import alignIO from './alignIO.ts';
 import { alignmentOf, customAlignmentOf } from './alignmentOf.ts';
 import type { AnyConcreteData, AnyData, Disarray, LooseDecorated, Unstruct } from './dataTypes.ts';
@@ -20,6 +26,10 @@ import {
   vec4u,
 } from './vector.ts';
 import type * as wgsl from './wgslTypes.ts';
+import type { BaseData } from './wgslTypes.ts';
+import type { BufferWriteOptions } from '../core/buffer/buffer.ts';
+import { getCompiledWriter } from './compiledIO.ts';
+import { getName } from '../shared/meta.ts';
 
 type DataWriter<TSchema extends wgsl.BaseData> = (
   output: ISerialOutput,
@@ -787,4 +797,61 @@ export function readData<TData extends wgsl.BaseData>(
   }
 
   return reader(input, schema);
+}
+
+const endianness = getSystemEndianness();
+
+export function writeToArrayBuffer<T extends BaseData>(
+  buffer: ArrayBuffer,
+  schema: T,
+  data: InferInput<T> | ArrayBuffer,
+  options?: BufferWriteOptions,
+) {
+  const startOffset = options?.startOffset ?? 0;
+  const endOffset = options?.endOffset ?? buffer.byteLength;
+
+  // Fast path: raw byte copy, user guarantees the padded layout
+  if (data instanceof ArrayBuffer || ArrayBuffer.isView(data)) {
+    const src =
+      data instanceof ArrayBuffer
+        ? new Uint8Array(data)
+        : new Uint8Array(data.buffer, data.byteOffset, data.byteLength);
+    const regionSize = endOffset - startOffset;
+    if (src.byteLength !== regionSize) {
+      console.warn(
+        `Buffer size mismatch: expected ${regionSize} bytes, got ${src.byteLength}. ` +
+          (src.byteLength < regionSize ? 'Data truncated.' : 'Excess ignored.'),
+      );
+    }
+    const copyLen = Math.min(src.byteLength, regionSize);
+    new Uint8Array(buffer).set(src.subarray(0, copyLen), startOffset);
+    return;
+  }
+
+  const dataView = new DataView(buffer);
+  const isLittleEndian = endianness === 'little';
+
+  const compiledWriter = getCompiledWriter(schema);
+
+  if (compiledWriter) {
+    try {
+      compiledWriter(dataView, startOffset, data, isLittleEndian, endOffset);
+      return;
+    } catch (error) {
+      console.error(
+        `Error when using compiled writer for data type '${
+          schema.type
+        }' (${getName(schema) ?? 'unnamed'}) - this is likely a bug, please submit an issue at https://github.com/software-mansion/TypeGPU/issues\nUsing fallback writer instead.`,
+        error,
+      );
+    }
+  }
+
+  const writer = new BufferWriter(buffer);
+  writer.seekTo(startOffset);
+  writeData(writer, schema, data as Infer<T>);
+}
+
+export function readFromArrayBuffer<T extends BaseData>(buffer: ArrayBuffer, schema: T): Infer<T> {
+  return readData(new BufferReader(buffer), schema);
 }
